pyproject.toml
[project]
name = "rag-workshop"
version = "0.1.0"
description = "RAG Workshop: Smart Course Advisor"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core LangChain
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-huggingface>=0.1.0",
    "langchain-chroma>=0.2.0",

    # Vector Store
    "chromadb>=0.5.0",

    # Embeddings
    "sentence-transformers>=3.0.0",

    # UI
    "streamlit>=1.40.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",

    # Data processing
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",

    # Evaluation (optional)
    "ragas>=0.2.0",

    # Jupyter for notebooks
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
